third_party/blink/web_tests/fast/css/getComputedStyle/getComputedStyle-border-style-shorthand-expected.txt

Tests that the border-style shorthand is computed properly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS computedStyle.getPropertyValue('border-style') is 'solid dotted'
PASS computedStyle.getPropertyValue('border-style') is 'solid dotted groove dashed'
PASS successfullyParsed is true

TEST COMPLETE
